fix: 7주차 미션 피드백 반영
