+++ week7/app/.idea/libraries/R2.xml
@@ -1,0 +1,9 @@
+<component name="libraryTable">
+  <library name="R2">
+    <CLASSES>
+      <root url="jar://$PROJECT_DIR$/build/intermediates/compile_and_runtime_r_class_jar/debug/processDebugResources/R.jar!/" />
+    </CLASSES>
+    <JAVADOC />
+    <SOURCES />
+  </library>
+</component>
