+++ week7/app/.idea/libraries/out.xml
@@ -1,0 +1,19 @@
+<component name="libraryTable">
+  <library name="out">
+    <CLASSES>
+      <root url="jar://$PROJECT_DIR$/build/intermediates/project_dex_archive/debug/dexBuilderDebug/out/e81ca0ffaa326fd29b41bd15812c635207f11e1bb02e2bac09ec538693424a5f_7.jar!/" />
+      <root url="jar://$PROJECT_DIR$/build/intermediates/project_dex_archive/debug/dexBuilderDebug/out/e81ca0ffaa326fd29b41bd15812c635207f11e1bb02e2bac09ec538693424a5f_8.jar!/" />
+      <root url="jar://$PROJECT_DIR$/build/intermediates/project_dex_archive/debug/dexBuilderDebug/out/e81ca0ffaa326fd29b41bd15812c635207f11e1bb02e2bac09ec538693424a5f_2.jar!/" />
+      <root url="jar://$PROJECT_DIR$/build/intermediates/project_dex_archive/debug/dexBuilderDebug/out/e81ca0ffaa326fd29b41bd15812c635207f11e1bb02e2bac09ec538693424a5f_9.jar!/" />
+      <root url="jar://$PROJECT_DIR$/build/intermediates/project_dex_archive/debug/dexBuilderDebug/out/e81ca0ffaa326fd29b41bd15812c635207f11e1bb02e2bac09ec538693424a5f_0.jar!/" />
+      <root url="jar://$PROJECT_DIR$/build/intermediates/project_dex_archive/debug/dexBuilderDebug/out/e81ca0ffaa326fd29b41bd15812c635207f11e1bb02e2bac09ec538693424a5f_5.jar!/" />
+      <root url="jar://$PROJECT_DIR$/build/intermediates/project_dex_archive/debug/dexBuilderDebug/out/e81ca0ffaa326fd29b41bd15812c635207f11e1bb02e2bac09ec538693424a5f_3.jar!/" />
+      <root url="jar://$PROJECT_DIR$/build/intermediates/project_dex_archive/debug/dexBuilderDebug/out/e81ca0ffaa326fd29b41bd15812c635207f11e1bb02e2bac09ec538693424a5f_10.jar!/" />
+      <root url="jar://$PROJECT_DIR$/build/intermediates/project_dex_archive/debug/dexBuilderDebug/out/e81ca0ffaa326fd29b41bd15812c635207f11e1bb02e2bac09ec538693424a5f_4.jar!/" />
+      <root url="jar://$PROJECT_DIR$/build/intermediates/project_dex_archive/debug/dexBuilderDebug/out/e81ca0ffaa326fd29b41bd15812c635207f11e1bb02e2bac09ec538693424a5f_1.jar!/" />
+      <root url="jar://$PROJECT_DIR$/build/intermediates/project_dex_archive/debug/dexBuilderDebug/out/e81ca0ffaa326fd29b41bd15812c635207f11e1bb02e2bac09ec538693424a5f_6.jar!/" />
+    </CLASSES>
+    <JAVADOC />
+    <SOURCES />
+  </library>
+</component>
