+++ week7/app/src/main/java/com/example/week7/ui/cart/CartScreen.kt
@@ -1,0 +1,51 @@
+package com.example.week7.ui.cart
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Button
+import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.example.week7.R
+
+@Composable
+fun CartScreen(onOrderClick: () -> Unit) {
+    Box(
+        modifier = Modifier.fillMaxSize(),
+        contentAlignment = Alignment.Center
+    ) {
+        Column(
+            horizontalAlignment = Alignment.CenterHorizontally
+        ) {
+            Image(
+                painter = painterResource(id = R.drawable.ic_bagcircle),
+                contentDescription = null,
+                modifier = Modifier.size(80.dp)
+            )
+            Spacer(modifier = Modifier.height(16.dp))
+            Text(text = "장바구니가 비어 있습니다.", fontWeight = FontWeight.Bold)
+            Text(text = "제품을 추가하면 여기에 표시됩니다.", color = Color.Gray, fontSize = 12.sp)
+        }
+
+        Button(
+            onClick = onOrderClick,
+            modifier = Modifier
+                .align(Alignment.BottomCenter)
+                .padding(bottom = 32.dp)
+                .fillMaxWidth(0.8f)
+                .height(56.dp),
+            colors = ButtonDefaults.buttonColors(containerColor = Color.Black),
+            shape = RoundedCornerShape(28.dp)
+        ) {
+            Text(text = "주문하기", color = Color.White)
+        }
+    }
+}
