+++ week7/app/src/main/java/com/example/week7/ui/shop/ShopScreen.kt
@@ -1,0 +1,32 @@
+package com.example.week7.ui.shop
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.*
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun ShopScreen() {
+    Column(
+        modifier = Modifier
+            .fillMaxSize()
+            .padding(16.dp)
+    ) {
+        Row(
+            modifier = Modifier.fillMaxWidth(),
+            horizontalArrangement = Arrangement.spacedBy(16.dp)
+        ) {
+            Column(horizontalAlignment = Alignment.CenterHorizontally) {
+                Text(text = "전체", fontWeight = FontWeight.Bold)
+                Box(modifier = Modifier.width(40.dp).height(2.dp).background(Color.Black))
+            }
+            Text(text = "TOPS & T-SHIRTS", color = Color.Gray)
+            Text(text = "SHOES", color = Color.Gray)
+        }
+    }
+}
