+++ week7/app/src/main/java/com/example/week7/ui/wishlist/WishlistScreen.kt
@@ -1,0 +1,20 @@
+package com.example.week7.ui.wishlist
+
+import androidx.compose.foundation.layout.*
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+
+@Composable
+fun WishlistScreen() {
+    Column(
+        modifier = Modifier
+            .fillMaxSize()
+            .padding(16.dp)
+    ) {
+        Text(text = "위시리스트", fontSize = 24.sp, fontWeight = FontWeight.Bold)
+    }
+}
